# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CSE 341 Project 2 API",
        "description": "REST API performing full CRUD operations on the contacts, users, colleges, courses, and instructors collections in MongoDB. All routes include data validation and error handling, and return proper HTTP status codes. Every POST, PUT, and DELETE route requires an authenticated GitHub session and returns 401 otherwise. To try a protected route in Swagger UI, open /login in a browser and sign in with the same host, then reload this page.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://web-servises-cse-341.onrender.com/",
            "description": "Production (Render)"
        },
        {
            "url": "http://localhost:3000",
            "description": "Local development"
        }
    ],
...
    "paths": {
...
        "/colleges": {
            "get": {
                "tags": ["Colleges"],
                "operationId": "getColleges",
                "summary": "Get all colleges",
                "description": "Returns a list of every college in the colleges collection.",
                "responses": {
                    "200": {
                        "description": "A list of colleges",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/College"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            },
            "post": {
                "tags": ["Colleges"],
                "operationId": "createCollege",
                "security": [{ "githubOAuth": [] }],
                "summary": "Create a college",
                "description": "Inserts a new college into the colleges collection. All fields are required.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/College"
                            },
                            "examples": {
                                "college": {
                                    "value": {
                                        "name": "Makerere University",
                                        "code": "MAK",
                                        "email": "admissions@mak.ac.ug",
                                        "phone": "+256-414-422-000",
                                        "city": "Kampala",
                                        "country": "Uganda",
                                        "establishedYear": 1922,
                                        "type": "public",
                                        "website": "https://www.mak.ac.ug"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "College created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "64f1e2d3a5b6c7d8e9f0a1b2"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing required fields or invalid data"
                    },
                    "401": {
                        "description": "Not authenticated; log in with GitHub first"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            }
        },
        "/colleges/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "The id of the college",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "get": {
                "tags": ["Colleges"],
                "operationId": "getCollegeById",
                "summary": "Get a college by id",
                "description": "Returns a single college matching the given id, including its courses.",
                "responses": {
                    "200": {
                        "description": "The requested college with its courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollegeWithCourses"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid college id"
                    },
                    "404": {
                        "description": "College not found"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            },
            "put": {
                "tags": ["Colleges"],
                "operationId": "updateCollege",
                "security": [{ "githubOAuth": [] }],
                "summary": "Update a college by id",
                "description": "Replaces the fields of the college matching the given id. All fields are required.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/College"
                            },
                            "examples": {
                                "college": {
                                    "value": {
                                        "name": "Makerere University",
                                        "code": "MAK",
                                        "email": "admissions.updated@mak.ac.ug",
                                        "phone": "+256-414-422-001",
                                        "city": "Kampala",
                                        "country": "Uganda",
                                        "establishedYear": 1922,
                                        "type": "public",
                                        "website": "https://www.mak.ac.ug"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "College updated"
                    },
                    "400": {
                        "description": "Invalid college id or missing fields"
                    },
                    "401": {
                        "description": "Not authenticated; log in with GitHub first"
                    },
                    "404": {
                        "description": "College not found"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            },
            "delete": {
                "tags": ["Colleges"],
                "operationId": "deleteCollege",
                "security": [{ "githubOAuth": [] }],
                "summary": "Delete a college by id",
                "description": "Deletes the college matching the given id, along with any courses that reference it.",
                "responses": {
                    "200": {
                        "description": "College and its courses deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "deletedColleges": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "deletedCourses": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid college id"
                    },
                    "401": {
                        "description": "Not authenticated; log in with GitHub first"
                    },
                    "404": {
                        "description": "College not found"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "githubOAuth": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://github.com/login/oauth/authorize",
                        "tokenUrl": "https://github.com/login/oauth/access_token",
                        "scopes": {
                            "user:email": "Read access to your email address"
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "College": {
                "type": "object",
                "required": ["name", "code", "email", "phone", "city", "country", "establishedYear", "type", "website"],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Makerere University"
                    },
                    "code": {
                        "type": "string",
                        "description": "Short unique code, 2-12 letters, numbers, or dashes.",
                        "example": "MAK"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admissions@mak.ac.ug"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+256-414-422-000"
                    },
                    "city": {
                        "type": "string",
                        "example": "Kampala"
                    },
                    "country": {
                        "type": "string",
                        "example": "Uganda"
                    },
                    "establishedYear": {
                        "type": "integer",
                        "description": "Four-digit year the college was established.",
                        "example": 1922
                    },
                    "type": {
                        "type": "string",
                        "enum": ["public", "private", "community", "technical"],
                        "example": "public"
                    },
                    "website": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://www.mak.ac.ug"
                    }
                }
            },
            "CollegeWithCourses": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/College"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "_id": {
                                "type": "string",
                                "example": "64f1e2d3a5b6c7d8e9f0a1b2"
                            },
                            "courses": {
                                "type": "array",
                                "description": "Courses that reference this college.",
                                "items": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    }
                ]
            },
            "Course": {
                "type": "object",
                "required": ["name", "code", "collegeId", "credits", "duration", "level", "faculty", "description"],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Bachelor of Science in Computer Science"
                    },
                    "code": {
                        "type": "string",
                        "description": "Short unique code, 2-12 letters, numbers, or dashes.",
                        "example": "BSCS"
                    },
                    "collegeId": {
                        "type": "string",
                        "description": "The _id of the college that offers this course. Must match an existing college.",
                        "example": "64f1e2d3a5b6c7d8e9f0a1b2"
                    },
                    "credits": {
                        "type": "integer",
                        "description": "Total credit hours for the course.",
                        "example": 120
                    },
                    "duration": {
                        "type": "integer",
                        "description": "Length of the course in years.",
                        "example": 4
                    },
                    "level": {
                        "type": "string",
                        "enum": ["certificate", "diploma", "bachelors", "masters", "phd"],
                        "example": "bachelors"
                    },
                    "faculty": {
                        "type": "string",
                        "example": "Faculty of Technology"
                    },
                    "description": {
                        "type": "string",
                        "example": "Covers programming, algorithms, databases, and software engineering."
                    }
                }
            },
...
        }
    }
}